+++ react/CHANGELOG.md
@@ -1,3 +1,11 @@
+### [1.0.0] - 2024-07-03
+
+-   Added Newsletter & Memberships support
+-   Updaded to @hyvor/hyvor-talk-base 1.0.0 new API
+-   Minor breaking changes
+    -   `translations` attribute in `Comments` component has new types
+    -   `<CommentCount>` no longer automatically loads the counts. Needs to call `CommentCounts.load()` manually from `@hyvor/hyvor-talk-base`
+
 ### [0.0.3] - 2024-03-10
 
 -   Updated base version to support newer types